% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{expect_snapshot_data}
\alias{expect_snapshot_data}
\title{Snapshot testing for [data.frame]s}
\usage{
expect_snapshot_data(x, name, digits = 6)
}
\arguments{
\item{x}{a [data.frame] to snapshot}

\item{name}{[character] snapshot name}

\item{digits}{[integer] passed to [signif()] for numeric variables}
}
\value{
[NULL] (from [testthat::expect_snapshot_file()])
}
\description{
Snapshot testing for [data.frame]s
}
\examples{
expect_snapshot_data(iris, name = iris)
}
